+++ packages/xchain-util/src/chain.const.ts
@@ -22,8 +22,16 @@ export const CosmosChain = 'GAIA'
  * Polkadot Chain
  */
 export const PolkadotChain = 'POLKA'
+/**
+ * Bitcoin Cash
+ */
+export const BCHChain = 'BCH'
+/**
+ * Litecoin Chain
+ */
+export const LTCChain = 'LTC'
 
 /**
  * All possible chains XChainJS currently supports
  * */
-export const chains = [BNBChain, BTCChain, ETHChain, THORChain, CosmosChain, PolkadotChain] as const
+export const chains = [BNBChain, BTCChain, ETHChain, THORChain, CosmosChain, PolkadotChain, BCHChain, LTCChain] as const